+++ packages/block-editor/src/components/block-list/use-block-props/index.js
@@ -8,23 +8,19 @@ import { omit } from 'lodash';
  * WordPress dependencies
  */
 import { useRef, useEffect, useContext } from '@wordpress/element';
-import { isTextField } from '@wordpress/dom';
-import { ENTER, BACKSPACE, DELETE } from '@wordpress/keycodes';
 import { __, sprintf } from '@wordpress/i18n';
-import { useDispatch } from '@wordpress/data';
 import { __unstableGetBlockProps as getBlockProps } from '@wordpress/blocks';
 
 /**
  * Internal dependencies
  */
-import { isInsideRootBlock } from '../../../utils/dom';
 import useMovingAnimation from '../../use-moving-animation';
 import { SetBlockNodes } from '../root-container';
-import { SelectionStart } from '../../writing-flow';
 import { BlockListBlockContext } from '../block';
 import { useFocusFirstElement } from './use-focus-first-element';
 import { useIsHovered } from './use-is-hovered';
 import { useBlockMovingModeClassNames } from './use-block-moving-mode-class-names';
+import { useEventHandlers } from './use-event-handlers';
 
 /**
  * This hook is used to lightly mark an element as a block element. The element
@@ -45,11 +41,9 @@ import { useBlockMovingModeClassNames } from './use-block-moving-mode-class-name
 export function useBlockProps( props = {}, { __unstableIsHtml } = {} ) {
 	const fallbackRef = useRef();
 	const ref = props.ref || fallbackRef;
-	const onSelectionStart = useContext( SelectionStart );
 	const setBlockNodes = useContext( SetBlockNodes );
 	const {
 		clientId,
-		rootClientId,
 		isSelected,
 		isFirstMultiSelected,
 		isLastMultiSelected,
@@ -62,9 +56,6 @@ export function useBlockProps( props = {}, { __unstableIsHtml } = {} ) {
 		blockTitle,
 		wrapperProps = {},
 	} = useContext( BlockListBlockContext );
-	const { insertDefaultBlock, removeBlock, selectBlock } = useDispatch(
-		'core/block-editor'
-	);
 
 	// Provide the selected node, or the first and last nodes of a multi-
 	// selection, so it can be used to position the contextual block toolbar.
@@ -101,6 +92,7 @@ export function useBlockProps( props = {}, { __unstableIsHtml } = {} ) {
 	const blockLabel = sprintf( __( 'Block: %s' ), blockTitle );
 
 	useFocusFirstElement( ref, clientId );
+	useEventHandlers( ref, clientId );
 
 	// Block Reordering animation
 	useMovingAnimation(
@@ -110,95 +102,6 @@ export function useBlockProps( props = {}, { __unstableIsHtml } = {} ) {
 		enableAnimation,
 		index
 	);
-
-	useEffect( () => {
-		if ( ! isSelected ) {
-			/**
-			 * Marks the block as selected when focused and not already
-			 * selected. This specifically handles the case where block does not
-			 * set focus on its own (via `setFocus`), typically if there is no
-			 * focusable input in the block.
-			 *
-			 * @param {FocusEvent} event Focus event.
-			 */
-			function onFocus( event ) {
-				// If an inner block is focussed, that block is resposible for
-				// setting the selected block.
-				if ( ! isInsideRootBlock( ref.current, event.target ) ) {
-					return;
-				}
-
-				selectBlock( clientId );
-			}
-
-			ref.current.addEventListener( 'focus', onFocus, true );
-
-			return () => {
-				ref.current.removeEventListener( 'focus', onFocus, true );
-			};
-		}
-
-		/**
-		 * Interprets keydown event intent to remove or insert after block if
-		 * key event occurs on wrapper node. This can occur when the block has
-		 * no text fields of its own, particularly after initial insertion, to
-		 * allow for easy deletion and continuous writing flow to add additional
-		 * content.
-		 *
-		 * @param {KeyboardEvent} event Keydown event.
-		 */
-		function onKeyDown( event ) {
-			const { keyCode, target } = event;
-
-			if (
-				keyCode !== ENTER &&
-				keyCode !== BACKSPACE &&
-				keyCode !== DELETE
-			) {
-				return;
-			}
-
-			if ( target !== ref.current || isTextField( target ) ) {
-				return;
-			}
-
-			event.preventDefault();
-
-			if ( keyCode === ENTER ) {
-				insertDefaultBlock( {}, rootClientId, index + 1 );
-			} else {
-				removeBlock( clientId );
-			}
-		}
-
-		function onMouseLeave( { buttons } ) {
-			// The primary button must be pressed to initiate selection.
-			// See https://developer.mozilla.org/en-US/docs/Web/API/MouseEvent/buttons
-			if ( buttons === 1 ) {
-				onSelectionStart( clientId );
-			}
-		}
-
-		/**
-		 * Prevents default dragging behavior within a block. To do: we must
-		 * handle this in the future and clean up the drag target.
-		 *
-		 * @param {DragEvent} event Drag event.
-		 */
-		function onDragStart( event ) {
-			event.preventDefault();
-		}
-
-		ref.current.addEventListener( 'keydown', onKeyDown );
-		ref.current.addEventListener( 'mouseleave', onMouseLeave );
-		ref.current.addEventListener( 'dragstart', onDragStart );
-
-		return () => {
-			ref.current.removeEventListener( 'mouseleave', onMouseLeave );
-			ref.current.removeEventListener( 'keydown', onKeyDown );
-			ref.current.removeEventListener( 'dragstart', onDragStart );
-		};
-	}, [ isSelected, onSelectionStart, insertDefaultBlock, removeBlock ] );
 
 	const isHovered = useIsHovered( ref );
 	const blockMovingModeClassNames = useBlockMovingModeClassNames( clientId );
